fix(LinstorSR): enabled/disable minidrbcluster with fixed order
Ensure we disable minidrbdcluster during SR destruction on all hosts
with a deterministic execution to ensure linstor-controller is never restarted
on another host. It was possible before this patch because the host of the minidrbcluster
that was running the controller could be stopped before the others. Now the primary service
is stopped last.

Signed-off-by: Ronan Abhamon <ronan.abhamon@vates.fr>